refactor: take out user functions
