+++ index.js
@@ -1,15 +1,6 @@
 import readline from 'readline';
 import {getCurrentDirectory, printCWD, setInitialDirectory} from "./src/utils/directory.js";
-
-function parseArgs() {
-  const cliArg = process.argv.find(arg => arg.startsWith('--username='));
-  const envArg = process.env.npm_config_username;
-  if (cliArg) return cliArg.split('=')[1];
-  if (envArg) return envArg;
-  return 'Anonymous';
-}
-
-const username = parseArgs();
+import {exitHandler, printWelcomeMessage} from "./src/utils/user.js";
 
 const rl = readline.createInterface({
   input: process.stdin,
@@ -17,18 +8,13 @@ const rl = readline.createInterface({
   prompt: '> ',
 });
 
-function exitHandler(username) {
-  console.log(`\nThank you for using File Manager, ${username}, goodbye!`);
-  process.exit(0);
-}
-
 function handleCommand(input, currentDir) {
 
 }
 
 setInitialDirectory();
 
-console.log(`Welcome to the File Manager, ${username}!`);
+printWelcomeMessage();
 printCWD(getCurrentDirectory());
 rl.prompt();
 
@@ -49,7 +35,6 @@ rl.on('line', async (input) => {
   }
 });
 
-
 rl
-.on('close', () => exitHandler(username))
-.on('SIGINT', () => exitHandler(username));
+.on('close', () => exitHandler())
+.on('SIGINT', () => exitHandler());
